feat: wire shared video capability registry into MCP planning
