+++ src/config/modelArchetypes/types.ts
@@ -10,14 +10,13 @@
 // 现有层只有按 key 名猜的 image-url，表达不了 character1..N / 视频 / 音频）。
 
 import type { ModelParameterControl } from "../modelCatalogMeta";
+import type {
+  ArchetypeExpressionChannel as SharedArchetypeExpressionChannel,
+  ArchetypeReferenceSlotKind as SharedArchetypeReferenceSlotKind,
+  ArchetypeSource as SharedArchetypeSource,
+} from "../../../electron/shared/videoCapabilities/types";
 
-export type ArchetypeReferenceSlotKind =
-  | "first_frame"
-  | "last_frame"
-  | "image_ref" // 多图，按序对应 prompt 里的 character1..N
-  | "video_ref"
-  | "audio_ref"
-  | "source_video";
+export type ArchetypeReferenceSlotKind = SharedArchetypeReferenceSlotKind;
 
 export type ArchetypeReferenceSlot = {
   kind: ArchetypeReferenceSlotKind;
@@ -70,15 +69,7 @@ export type ArchetypeReferenceSlot = {
  * `status` distinguishes documented support from a documented rejection and from
  * an area that has not been verified yet.
  */
-export type ArchetypeExpressionChannel = {
-  signal: string;
-  via: "prompt" | "reference_slot" | "structured_parameter";
-  status: "documented" | "unsupported" | "unknown";
-  slotKind?: ArchetypeReferenceSlotKind;
-  parameterKey?: string;
-  parameterPath?: string;
-  evidence?: ArchetypeSource;
-};
+export type ArchetypeExpressionChannel = SharedArchetypeExpressionChannel;
 
 /** 跨模型统一的「意图」——UI 主标签按它走（角色参考/单图首帧/首尾帧/文生/视频编辑）。 */
 export type ArchetypeIntent = "text" | "single" | "firstlast" | "character" | "edit";
@@ -180,16 +171,7 @@ export type ModelArchetypeVariant = {
  * 门岗检查、被聚合成索引、让复核的人一眼知道去哪查。**改契约数字必须同步 checkedAt。**
  * 门岗：`pnpm run check:archetype-sources`（棘轮，只减不增）。
  */
-export type ArchetypeSource = {
-  /** 供应商**官方**文档 URL（不是第三方转述——转述会漏会错，栽过）。 */
-  url: string;
-  /** 最后一次逐项对账的日期，ISO `YYYY-MM-DD`。 */
-  checkedAt: string;
-  /** 该文档属于哪个供应商；模型级、跨供应商通用的可省略。 */
-  vendorKey?: string;
-  /** 一句话：这份文档覆盖了什么（如「参考通道字段名与上限」）。 */
-  covers?: string;
-};
+export type ArchetypeSource = SharedArchetypeSource;
 
 export type ModelArchetype = {
   id: string; // 'seedance-2'
